requirements.txt
pyyaml
docker
markdown
sh
